feat(api):: Move PlayerStats interface to a separate file
