+++ src/Frontend/code/api/leaderboard/getTopTenAPI.ts
@@ -1,12 +1,4 @@
-export interface PlayerStats {
-  id: number; // userID
-  username: string;
-  rank?: number; // Rank of the player in the leaderboar
-  wins: number;
-  losses: number;
-  points: number; // Points of the player in the leaderboard
-}
-
+import { PlayerStats } from "./types/PlayerStatsInterface";
 /**
  * @brief
  *
