+++ src/Frontend/code/api/leaderboard/types/PlayerStatsInterface.ts
@@ -1,0 +1,8 @@
+export interface PlayerStats {
+  id: number; // userID
+  username: string;
+  rank?: number; // Rank of the player in the leaderboar
+  wins: number;
+  losses: number;
+  points: number; // Points of the player in the leaderboard
+}
